apiserver/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.94
botocore==1.34.94
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django>=5.1.0
django-environ==0.11.2
django-extensions==3.2.3
django-ninja==1.1.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
jmespath==1.0.1
kombu==5.3.7
packaging==24.0
pika==1.3.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
s3transfer==0.10.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.3
pymysql==1.1.1
